feat: merge FlowDesigner.Web into FlowDesigner.Api
- Integrate Blazor Server into ASP.NET Core API project
- Add RazorPages, ServerSideBlazor services and middleware
- Create _Host.cshtml for Blazor Server entry point
- Move all Web project files (Pages, Components, Services, wwwroot) to Api
- Update namespaces from FlowDesigner.Web to FlowDesigner.Api
- Register FlowApiService with HttpClient via DI
- Remove FlowDesigner.Web project from solution
- Fix compilation errors across multiple service files
- Add .gitignore for bin/obj directories

+++ src/FlowDesigner.Api/Services/RtpService.cs
@@ -62,7 +62,7 @@ public class RtpService
             if (config.Multicast)
             {
                 client.JoinMulticastGroup(IPAddress.Parse(config.Host));
-                client.Ttl = config.Ttl;
+                client.Ttl = (short)config.Ttl;
             }
 
             if (!config.IsSender)
